feat(celo-transfers): index CELO transfers (#28)
relies on #26 and #27

## Testing

- unit test coverage for handleTransfers
- ran locally for a while and indexed some celo transfers, looked fine.

## Caveats on data incompleteness

there is some weirdness to this, because you can actually transfer CELO in 2 ways:
- by interacting with the ERC20 contract (which will emit events that are indexed)
- by specifying a "value" field in a transaction

reference [here](https://docs.celo.org/developer/migrate/from-ethereum#:~:text=CELO%20has%20an%20ERC20%20interface,go%20through%20the%20token%20contract.)

So we need to be careful in the assumptions we make about what this data really represents, since it's incomplete. 

I think we make the assumption elsewhere-- like in blockchain-api-- that the smart contract CELO transfers are the only ones to worry about, so my hope is that the vast majority of CELO transfers use the smart contract. But I don't have data to support that at this time.

I suggest we merge this and just be careful about data consumption until the question is resolved. Created a Linear issue for further investigation [here](https://linear.app/valora/issue/ACT-566/spike-celo-transfer-data-incompleteness)